+++ insteon_mqtt/device/Dimmer.py
@@ -46,7 +46,7 @@ class Dimmer(Base):
 
     def __init__(self, protocol, modem, address, name=None):
         """Constructor
-
+        
         Args:
           protocol (Protocol): The Protocol object used to communicate
                    with the Insteon network.  This is needed to allow the
@@ -522,6 +522,41 @@ class Dimmer(Base):
         self.send(msg, msg_handler)
 
     #-----------------------------------------------------------------------
+    def set_ramp_rate(self, rate, on_done=None):
+        """Set the device default ramp rate.
+
+        This changes the dimmer default ramp rate of how quickly it will 
+        turn on or off. This rate can be between 0.1 seconds and up to 9 
+        minutes.
+
+        Args:
+          rate (float): Ramp rate in in the range [0.1, 540] seconds
+          on_done: Finished callback.  This is called when the command has
+                   completed.  Signature is: on_done(success, msg, data)
+        """
+        LOG.info("Dimmer %s setting ramp rate to %s", self.label, rate)
+
+        data_3 = 0x1c #the default ramp rate is .5
+        for ramp_key, ramp_value in self.ramp_pretty.items():
+            if rate >= ramp_value:
+                data_3 = ramp_key
+                break
+
+        # Extended message data - see Insteon dev guide p156.
+        data = bytes([
+            0x01,   # D1 must be group 0x01
+            0x05,   # D2 set ramp rate when button is pressed
+            data_3,  # D3 rate
+            ] + [0x00] * 11)
+
+        msg = Msg.OutExtended.direct(self.addr, 0x2e, 0x00, data)
+
+        # Use the standard command handler which will notify us when the
+        # command is ACK'ed.
+        msg_handler = handler.StandardCmd(msg, self.handle_ramp_rate, on_done)
+        self.send(msg, msg_handler)
+
+    #-----------------------------------------------------------------------
     def set_flags(self, on_done, **kwargs):
         """Set internal device flags.
 
@@ -545,7 +580,8 @@ class Dimmer(Base):
         # passed in.
         FLAG_BACKLIGHT = "backlight"
         FLAG_ON_LEVEL = "on_level"
-        flags = set([FLAG_BACKLIGHT, FLAG_ON_LEVEL])
+        FLAG_RAMP_RATE = "ramp_rate"
+        flags = set([FLAG_BACKLIGHT, FLAG_ON_LEVEL, FLAG_RAMP_RATE])
         unknown = set(kwargs.keys()).difference(flags)
         if unknown:
             raise Exception("Unknown Dimmer flags input: %s.\n Valid flags "
@@ -561,6 +597,10 @@ class Dimmer(Base):
         if FLAG_ON_LEVEL in kwargs:
             on_level = util.input_byte(kwargs, FLAG_ON_LEVEL)
             seq.add(self.set_on_level, on_level)
+
+        if FLAG_RAMP_RATE in kwargs:
+            rate = util.input_float(kwargs, FLAG_RAMP_RATE)
+            seq.add(self.set_ramp_rate, rate)
 
         seq.run()
 
@@ -599,6 +639,24 @@ class Dimmer(Base):
             on_done(True, "Button on level updated", None)
         else:
             on_done(False, "Button on level failed", None)
+
+    #-----------------------------------------------------------------------
+    def handle_ramp_rate(self, msg, on_done):
+        """Callback for handling set_ramp_rate() responses.
+
+        This is called when we get a response to the set_ramp_rate() command.
+        We don't need to do anything - just call the on_done callback with
+        the status.
+
+        Args:
+          msg (InpStandard): The response message from the command.
+          on_done: Finished callback.  This is called when the command has
+                   completed.  Signature is: on_done(success, msg, data)
+        """
+        if msg.flags.type == Msg.Flags.Type.DIRECT_ACK:
+            on_done(True, "Button ramp rate updated", None)
+        else:
+            on_done(False, "Button ramp rate failed", None)
 
     #-----------------------------------------------------------------------
     def handle_broadcast(self, msg):