+++ HISTORY.md
@@ -35,6 +35,9 @@
 - Device names are now printed when printing the database.  This makes reading
   the database output much easier.  ([PR 239][P239])
 
+- Added ability to set the default ramp rate of a dimmer using the ramp_rate
+  flag.  (thanks @jordanrounds)([PR 235][P235])
+
 ### Fixes
 
 - Database delta is updated on database writes.  This eliminates a number of
@@ -448,3 +451,4 @@ will add new features.
 [P219]: https://github.com/TD22057/insteon-mqtt/pull/219
 [P239]: https://github.com/TD22057/insteon-mqtt/pull/239
 [P244]: https://github.com/TD22057/insteon-mqtt/pull/244
+[P235]: https://github.com/TD22057/insteon-mqtt/pull/235